move common tests
