+++ Wrappers/Python/test/test_PluginsAstra_FBP.py
@@ -16,7 +16,7 @@
 #   limitations under the License.
 
 
-from cil.utilities.projector_tests import TestCommon_FBP_SIM
+from utils_projectors import TestCommon_FBP_SIM
 import unittest
 from utils import has_gpu_astra, has_astra
